docs/requirements.txt
Jinja2>=2.7
python-gflags>=2.0
